Create code
This function uses the fetch() method to make a request to the server and retrieve data. It then converts the response to JSON and returns the speedometer reading from the data object. Note that this function does not modify any external state and does not have any side effects, making it a pure function.